requirements.txt
requests
openai
python-dotenv
colorama
